스타일: 함수 분리, 생성자 이름 변경, 프로퍼티 이름 변경 (DocumentPath -> imageDirPath)
